improvement: created Language Selector component
